Added device descriptors for memory objects.
Memory objects now embed a device dependent descriptor for each device
in the memory object context. The goal of the descriptor is to track and
allow the interaction with the actual storage for a given device. This
is needed to implement the coherency protocol of memory objects among
all the devices in the current context.