Ensure only one TargetMachine instance per device.
Avoid to create a TargetMachine within the optimizer each time we
compile a cl_program. For each device we have one target machine, when
optimizing a module for a given device, the target machine option must
be reset to match the optimization flags. The same must occur when we
generate machine code.