Fix: Built-in library
The "is_normal" function didn't check for a null argument, while
"select" definition needs to distinguish among scalar and vector
types and, for the latters, it must check the MSB (Most Significant
Bit) of each element of the last function parameter.

The "normalize" function now calls "length" instead of "rsqrt" and
"length" is based on the library "sqrt" function instead of the "sqrt"
and "sqrtf" from libc.

For the integer functions, some explicit casts have been added and the
template strategy has replace the recursion in some cases.