Fix: Asynchronous copy and prefetch built-ins
An inner loop inside AsyncCopyThread::Run() started counting from 1
intead of 0. The prefetch built-in implementation now starts fetching
from the end of the gentype buffer so that the beginning of the buffer
will end up higher in the cache hierarchy. It is likely that this will
be the first address referenced by subsequent code.