Fix: Memory coherency check and mapping flags
The CL_MAP_WRITE_INVALIDATE_REGION flag was not covered by the memory
coherency check for mapping commands. So far it is treated in the same
way as CL_MAP_WRITE but it should provide performance benefits over it
as it shouldn't guarantee that the latest bits are being mapped.